Eliminate encoding problems; switch to md5
Encoding
---

The encoding errors come when we try to convert to a string, but it
doesn't look like we need to convert the bytes into a string at any
point - we should simply hash the bytes.

I also took the liberty of splitting the file into 128KB chunks, to
avoid hogging memory.

MD5 over SHA1
---

Basically, all MD* and SH* hashing functions operate in much the same
way, but with different levels of complexity.

It is well-known by now that MD5's complexity is not great enough to
offer any meaningful security. Although SHA1 is significantly more
complex, it is also breakable in this day and age. Therefore, if we
wanted security, we should use at least SHA256.

However, we don't want security at all in this case, we want speed. [Some will say](https://stackoverflow.com/a/3431835/613540) that you should always use secure hashes "just in case", but I
disagree in general (use the function you need, insecurity comes from not understanding your needs properly) and especially since we've had speed
problems in this module in the past, I think we should optimise for
speed *rather than* security.

In this case, going for the faster MD5 function helps us not only be as
fast as possible, but also hopefully indicate more explicitly that
we're not offering security, as MD5's insecurity is quite well known
nowadays.