Rename to canonicalwebteam.versioned-static
We are now going to try to namespace all things under
`canonicalwebteam`.

Following [the latest
advice](https://packaging.python.org/namespace_packages/), I've tried to
make `canonicalwebteam` a "proper" namespace as much as possible, by
using pkgutil.extend_path.

In the near future, all our properties will be on Python 3.3, and so we
can use Native namespace packages. But that day is, unfortunately, not
today.